reduce duplicated code and create script to demonstrate alloc vars
