+++ tests/scripts/lang_features/report_and_alloc_vars.mm
@@ -1,0 +1,8 @@
+wasm:opcode:call:before {
+    alloc i32 count;
+    report i32 rep_count;
+
+    count++;
+
+    rep_count = count;
+}
